+++ client/src/components/list/ListBar.tsx
@@ -16,6 +16,7 @@
  * limitations under the License.
  */
 
+import cx from "classnames";
 import { Link } from "react-router-dom";
 import { EntityType } from "../../features/kgSearch";
 import SessionButton from "../../features/session/components/SessionButton";
@@ -25,7 +26,9 @@ import EntityCreators from "../entities/Creators";
 import EntityDescription from "../entities/Description";
 import EntityLabel from "../entities/Label";
 import VisibilityIcon from "../entities/VisibilityIcon";
-import { ListElementProps } from "./List.d";
+import PinnedBadge from "./PinnedBadge";
+import { ListElementProps } from "./list.types";
+
 import "./ListBar.scss";
 
 export function getMainActionByEntity(
@@ -64,7 +67,13 @@ function ListBar({
 
   return (
     <div className="container-entity-listBar">
-      <div className="entity-image">
+      <div className={cx("entity-image", "position-relative")}>
+        {/* <div
+          style={{ position: "absolute", top: 0, left: 0, right: 0, bottom: 0 }}
+        >
+          <PinFill />
+        </div> */}
+        <PinnedBadge entityType={itemType} slug={slug} />
         <Link to={url} className="text-decoration-none">
           <div
             style={imageStyles}
