+++ client/src/utils/helpers/KgSearchFunctions.ts
@@ -21,7 +21,7 @@ import {
   KgSearchResult,
   KgSearchResultLink,
 } from "../../features/kgSearch";
-import { ListElementProps } from "../../components/list/List.d";
+import { ListElementProps } from "../../components/list/list.types";
 import { Url } from "./url";
 import { DateFilterTypes } from "../../components/dateFilter/DateFilter";
 import { isEqual } from "lodash";
